pmi/pmi1: fix incompatibility with slurm's pmi2 PMI plugin
Slurm supports a maximum length of 1024 in a single command.
However it reports that the maximum supported value is 1024,
which is not possible when the maximum length of a command
is 1024.

As a workaround, if the MPI program is directly launched with slurm's
`srun --mpi=pmi2`, limit the maximum length of value to 512.

Signed-off-by: Yan-Jie Wang <yanjiewtw@gmail.com>